+++ C++/README.md
@@ -20,7 +20,7 @@ Implementation of well-known (and some rare) algorithms, in C++.
     * [Extended Euclidean algorithm](include/algorithm/number_theory/extended_euclidean.hpp) (Bézout coefficients) :white_check_mark:
   * [Fast exponentiation](source/number_theory/fast_exponentiation.cpp) :white_check_mark:
   * Nth Fibonacci number
-    * [Linear time algorithm](source/number_theory/fibonacci.cpp) :white_check_mark:
+    * [Linear time algorithm](include/algorithm/number_theory/fibonacci.hpp) :white_check_mark:
     * [Logarithmic time algorithm](source/number_theory/fibonacci_efficient.cpp) :white_check_mark:
   * [Perfect number check](source/number_theory/perfect_number_check.cpp) :white_check_mark:
   * Prime numbers
